% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devtools.R
\name{devtools}
\alias{devtools}
\alias{install}
\alias{load_all}
\alias{document}
\alias{check}
\title{devtool functions}
\usage{
install(...)

load_all(...)

document(...)

check(...)
}
\arguments{
\item{...}{Arguments passed to devtools functions}
}
\description{
Functions from devtools which will unpack/pack before and after execution.
See the devtools documentation for each of them.
}
\details{
devtools
}
