% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-separators.R
\name{\%,\%<-}
\alias{\%,\%<-}
\alias{\%,*\%<-}
\title{'Guide rail' exports}
\usage{
\%,\%(lhs, rhs) <- value

\%,*\%(lhs, rhs) <- value
}
\arguments{
\item{lhs}{Argument not intended to be used}

\item{rhs}{Argument not intended to be used}

\item{value}{Argument not intended to be used}
}
\description{
\code{packr} exports \code{\%,\%<-} and \code{\%,*\%<-} not because it uses these functions, but because it \emph{never} does. Due to the way \code{unpackr} evaluates lefthand sides of assignments, it never actually evaluates either of these functions, and if for some reason this is happening, something has gone wrong. These functions are exported to serve as 'guide rails' and alert the user that there is a problem.
}
