% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack_-_unpack.R
\name{unpack}
\alias{unpack}
\title{Unpack}
\usage{
unpack()
}
\description{
Unpacks all files in subdirectories of the R folder to conform to
R's flat file requirements. Make sure you `pack()` after doing this.
}
\details{
unpack
}
